# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rainbow Loom Vibe Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ColorAvailabilityColor = typeof ColorAvailabilityColor[keyof typeof ColorAvailabilityColor];


export const ColorAvailabilityColor = {
  red: 'red',
  orange: 'orange',
  green: 'green',
  blue: 'blue',
  purple: 'purple',
  pink: 'pink',
  black: 'black',
} as const;

export interface ColorAvailability {
  color: ColorAvailabilityColor;
  available: boolean;
}

export type CartItemPattern = typeof CartItemPattern[keyof typeof CartItemPattern];


export const CartItemPattern = {
  regular: 'regular',
  dragon_scale: 'dragon_scale',
  fish_scale: 'fish_scale',
  double_fish_scale: 'double_fish_scale',
} as const;

export interface Product {
  id: string;
  name: string;
  price: number;
  available: boolean;
  colors: ColorAvailability[];
}

export interface CartItem {
  productId: string;
  productName: string;
  colors: string[];
  pattern: CartItemPattern;
  price: number;
}

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  items: CartItem[];
}

export interface Order {
  id: string;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  items: CartItem[];
  total: number;
  status: string;
  estimatedCompletion?: string;
  createdAt: string;
}

export interface ColorAvailabilityUpdate {
  colors: ColorAvailability[];
}

export interface ProductAvailableUpdate {
  available: boolean;
}

export type GlobalColorUpdateColor = typeof GlobalColorUpdateColor[keyof typeof GlobalColorUpdateColor];


export const GlobalColorUpdateColor = {
  red: 'red',
  orange: 'orange',
  green: 'green',
  blue: 'blue',
  purple: 'purple',
  pink: 'pink',
  black: 'black',
} as const;

export interface GlobalColorUpdate {
  color: GlobalColorUpdateColor;
  available: boolean;
}

export interface StartOrderBody {
  days: number;
  hours: number;
  minutes: number;
}

export interface AdminCredentials {
  passcode: string;
}

export interface AdminToken {
  success: boolean;
  token: string;
}

export interface SuccessResult {
  success: boolean;
}

export interface ErrorResult {
  error: string;
}