# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rainbow Loom Vibe Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product and color availability operations
  - name: orders
    description: Order management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products with their color availability
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{productId}/colors:
    patch:
      operationId: updateProductColors
      tags: [products]
      summary: Admin - update color availability for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ColorAvailabilityUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /products/{productId}/available:
    patch:
      operationId: updateProductAvailable
      tags: [products]
      summary: Admin - toggle a product available/unavailable
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductAvailableUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /colors/global:
    patch:
      operationId: updateGlobalColor
      tags: [products]
      summary: Admin - enable or disable a color across ALL products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GlobalColorUpdate"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResult"

  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: Admin - get all orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order by ID
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /orders/{orderId}/hold:
    post:
      operationId: holdOrder
      tags: [orders]
      summary: Admin - put an order on hold and email customer
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order put on hold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /orders/{orderId}/unhold:
    post:
      operationId: unholdOrder
      tags: [orders]
      summary: Admin - remove hold and email customer their order is back in progress
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order unholded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /orders/{orderId}/start:
    post:
      operationId: startOrder
      tags: [orders]
      summary: Admin - start making an order with time estimate, email customer
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartOrderBody"
      responses:
        "200":
          description: Order started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [orders]
      summary: Admin login with passcode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToken"
        "401":
          description: Wrong passcode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ColorAvailability:
      type: object
      properties:
        color:
          type: string
          enum: [red, orange, green, blue, purple, pink, black]
        available:
          type: boolean
      required:
        - color
        - available

    CartItemPattern:
      type: string
      enum: [regular, dragon_scale, fish_scale, double_fish_scale]

    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        price:
          type: number
        available:
          type: boolean
        colors:
          type: array
          items:
            $ref: "#/components/schemas/ColorAvailability"
      required:
        - id
        - name
        - price
        - available
        - colors

    CartItem:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        colors:
          type: array
          items:
            type: string
        pattern:
          $ref: "#/components/schemas/CartItemPattern"
        price:
          type: number
      required:
        - productId
        - productName
        - colors
        - pattern
        - price

    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
      required:
        - customerName
        - customerEmail
        - items

    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        status:
          type: string
        estimatedCompletion:
          type: string
        createdAt:
          type: string
      required:
        - id
        - orderNumber
        - customerName
        - customerEmail
        - items
        - total
        - status
        - createdAt

    ColorAvailabilityUpdate:
      type: object
      properties:
        colors:
          type: array
          items:
            $ref: "#/components/schemas/ColorAvailability"
      required:
        - colors

    ProductAvailableUpdate:
      type: object
      properties:
        available:
          type: boolean
      required:
        - available

    GlobalColorUpdate:
      type: object
      properties:
        color:
          type: string
          enum: [red, orange, green, blue, purple, pink, black]
        available:
          type: boolean
      required:
        - color
        - available

    StartOrderBody:
      type: object
      properties:
        days:
          type: integer
        hours:
          type: integer
        minutes:
          type: integer
      required:
        - days
        - hours
        - minutes

    AdminCredentials:
      type: object
      properties:
        passcode:
          type: string
      required:
        - passcode

    AdminToken:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required:
        - success
        - token

    SuccessResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    ErrorResult:
      type: object
      properties:
        error:
          type: string
      required:
        - error