# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rainbow Loom Vibe Store API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Get all products with their color availability
 */
export const GetProductsResponseItem = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "price": zod.number(),
  "available": zod.boolean(),
  "colors": zod.array(zod.object({
  "color": zod.enum(['red', 'orange', 'green', 'blue', 'purple', 'pink', 'black']),
  "available": zod.boolean()
}))
})
export const GetProductsResponse = zod.array(GetProductsResponseItem)


/**
 * @summary Admin - update color availability for a product
 */
export const UpdateProductColorsParams = zod.object({
  "productId": zod.coerce.string()
})

export const UpdateProductColorsBody = zod.object({
  "colors": zod.array(zod.object({
  "color": zod.enum(['red', 'orange', 'green', 'blue', 'purple', 'pink', 'black']),
  "available": zod.boolean()
}))
})

export const UpdateProductColorsResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "price": zod.number(),
  "available": zod.boolean(),
  "colors": zod.array(zod.object({
  "color": zod.enum(['red', 'orange', 'green', 'blue', 'purple', 'pink', 'black']),
  "available": zod.boolean()
}))
})


/**
 * @summary Admin - toggle a product available/unavailable
 */
export const UpdateProductAvailableParams = zod.object({
  "productId": zod.coerce.string()
})

export const UpdateProductAvailableBody = zod.object({
  "available": zod.boolean()
})

export const UpdateProductAvailableResponse = zod.object({
  "id": zod.string(),
  "name": zod.string(),
  "price": zod.number(),
  "available": zod.boolean(),
  "colors": zod.array(zod.object({
  "color": zod.enum(['red', 'orange', 'green', 'blue', 'purple', 'pink', 'black']),
  "available": zod.boolean()
}))
})


/**
 * @summary Admin - enable or disable a color across ALL products
 */
export const UpdateGlobalColorBody = zod.object({
  "color": zod.enum(['red', 'orange', 'green', 'blue', 'purple', 'pink', 'black']),
  "available": zod.boolean()
})

export const UpdateGlobalColorResponse = zod.object({
  "success": zod.boolean()
})


/**
 * @summary Admin - get all orders
 */
export const GetOrdersResponseItem = zod.object({
  "id": zod.string(),
  "orderNumber": zod.string(),
  "customerName": zod.string(),
  "customerEmail": zod.string(),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "productName": zod.string(),
  "colors": zod.array(zod.string()),
  "pattern": zod.enum(['regular', 'dragon_scale', 'fish_scale', 'double_fish_scale']),
  "price": zod.number()
})),
  "total": zod.number(),
  "status": zod.string(),
  "estimatedCompletion": zod.string().optional(),
  "createdAt": zod.string()
})
export const GetOrdersResponse = zod.array(GetOrdersResponseItem)


/**
 * @summary Place a new order
 */
export const CreateOrderBody = zod.object({
  "customerName": zod.string(),
  "customerEmail": zod.string(),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "productName": zod.string(),
  "colors": zod.array(zod.string()),
  "pattern": zod.enum(['regular', 'dragon_scale', 'fish_scale', 'double_fish_scale']),
  "price": zod.number()
}))
})

export const CreateOrderResponse = zod.object({
  "id": zod.string(),
  "orderNumber": zod.string(),
  "customerName": zod.string(),
  "customerEmail": zod.string(),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "productName": zod.string(),
  "colors": zod.array(zod.string()),
  "pattern": zod.enum(['regular', 'dragon_scale', 'fish_scale', 'double_fish_scale']),
  "price": zod.number()
})),
  "total": zod.number(),
  "status": zod.string(),
  "estimatedCompletion": zod.string().optional(),
  "createdAt": zod.string()
})


/**
 * @summary Get a single order by ID
 */
export const GetOrderParams = zod.object({
  "orderId": zod.coerce.string()
})

export const GetOrderResponse = zod.object({
  "id": zod.string(),
  "orderNumber": zod.string(),
  "customerName": zod.string(),
  "customerEmail": zod.string(),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "productName": zod.string(),
  "colors": zod.array(zod.string()),
  "pattern": zod.enum(['regular', 'dragon_scale', 'fish_scale', 'double_fish_scale']),
  "price": zod.number()
})),
  "total": zod.number(),
  "status": zod.string(),
  "estimatedCompletion": zod.string().optional(),
  "createdAt": zod.string()
})


/**
 * @summary Admin - put an order on hold and email customer
 */
export const HoldOrderParams = zod.object({
  "orderId": zod.coerce.string()
})

export const HoldOrderResponse = zod.object({
  "id": zod.string(),
  "orderNumber": zod.string(),
  "customerName": zod.string(),
  "customerEmail": zod.string(),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "productName": zod.string(),
  "colors": zod.array(zod.string()),
  "pattern": zod.enum(['regular', 'dragon_scale', 'fish_scale', 'double_fish_scale']),
  "price": zod.number()
})),
  "total": zod.number(),
  "status": zod.string(),
  "estimatedCompletion": zod.string().optional(),
  "createdAt": zod.string()
})


/**
 * @summary Admin - remove hold and email customer their order is back in progress
 */
export const UnholdOrderParams = zod.object({
  "orderId": zod.coerce.string()
})

export const UnholdOrderResponse = zod.object({
  "id": zod.string(),
  "orderNumber": zod.string(),
  "customerName": zod.string(),
  "customerEmail": zod.string(),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "productName": zod.string(),
  "colors": zod.array(zod.string()),
  "pattern": zod.enum(['regular', 'dragon_scale', 'fish_scale', 'double_fish_scale']),
  "price": zod.number()
})),
  "total": zod.number(),
  "status": zod.string(),
  "estimatedCompletion": zod.string().optional(),
  "createdAt": zod.string()
})


/**
 * @summary Admin - start making an order with time estimate, email customer
 */
export const StartOrderParams = zod.object({
  "orderId": zod.coerce.string()
})

export const StartOrderBody = zod.object({
  "days": zod.number(),
  "hours": zod.number(),
  "minutes": zod.number()
})

export const StartOrderResponse = zod.object({
  "id": zod.string(),
  "orderNumber": zod.string(),
  "customerName": zod.string(),
  "customerEmail": zod.string(),
  "items": zod.array(zod.object({
  "productId": zod.string(),
  "productName": zod.string(),
  "colors": zod.array(zod.string()),
  "pattern": zod.enum(['regular', 'dragon_scale', 'fish_scale', 'double_fish_scale']),
  "price": zod.number()
})),
  "total": zod.number(),
  "status": zod.string(),
  "estimatedCompletion": zod.string().optional(),
  "createdAt": zod.string()
})


/**
 * @summary Admin login with passcode
 */
export const AdminLoginBody = zod.object({
  "passcode": zod.string()
})

export const AdminLoginResponse = zod.object({
  "success": zod.boolean(),
  "token": zod.string()
})