# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rainbow Loom Vibe Store API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItemPattern } from './cartItemPattern';

export interface CartItem {
  productId: string;
  productName: string;
  colors: string[];
  pattern: CartItemPattern;
  price: number;
}