# lib/api-zod/src/generated/types/cartItemPattern.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rainbow Loom Vibe Store API
 * OpenAPI spec version: 0.1.0
 */

export type CartItemPattern = typeof CartItemPattern[keyof typeof CartItemPattern];


export const CartItemPattern = {
  regular: 'regular',
  dragon_scale: 'dragon_scale',
  fish_scale: 'fish_scale',
  double_fish_scale: 'double_fish_scale',
} as const;