# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rainbow Loom Vibe Store API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface Order {
  id: string;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  items: CartItem[];
  total: number;
  status: string;
  estimatedCompletion?: string;
  createdAt: string;
}