# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rainbow Loom Vibe Store API
 * OpenAPI spec version: 0.1.0
 */
import type { CartItem } from './cartItem';

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  items: CartItem[];
}