# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rainbow Loom Vibe Store API
 * OpenAPI spec version: 0.1.0
 */
import type { ColorAvailability } from './colorAvailability';

export interface Product {
  id: string;
  name: string;
  price: number;
  available: boolean;
  colors: ColorAvailability[];
}