# lib/api-zod/src/generated/types/productAvailableUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Rainbow Loom Vibe Store API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductAvailableUpdate {
  available: boolean;
}